requirements.txt
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
cryptography==3.4.6
decorator==4.4.2
docker-pycreds==0.4.0
filelock==3.0.12
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
joblib==1.0.1
mkl-fft
mkl-random
mkl-service==2.3.0
networkx==2.5
numpy==1.19.2
olefile==0.46
packaging==20.9
pandas==1.2.3
pathtools==0.1.2
Pillow==8.1.2
prettytable==2.1.0
promise==2.3
protobuf==3.15.6
psutil==5.8.0
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
torch==1.7.0
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.0
sentry-sdk==1.0.0
setuptools
shortuuid==1.0.1
six==1.15.0
smmap==3.0.5
subprocess32==3.5.4
threadpoolctl==2.1.0
tokenizers==0.10.1
tqdm==4.59.0
transformers==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.4
wandb==0.10.22
wcwidth==0.2.5
wheel==0.36.2
onnx==1.8.1
onnxruntime-gpu==1.7.0
ai-hub
flask